Rename editor and input to more actively reflect their purpose
