+++ public/video-ui/src/components/FormFields/RichTextEditor.tsx
@@ -1,164 +1,77 @@
-import React from 'react';
-import RequiredForComposer from '../../constants/requiredForComposer';
+import React, { useRef, useEffect, useState } from 'react';
+import { EditorView } from 'prosemirror-view';
+import { createEditorView } from './richtext/setup';
+import { createSchema } from './richtext/createSchema';
+import { AllSelection, Transaction } from 'prosemirror-state';
+import { DOMParser as parseDOM, Node } from 'prosemirror-model';
 import { EditorConfig } from './richtext/config';
-import { getWords, isTooLong } from './richtext/utils/richTextHelpers';
-import { RichTextInput } from './RichTextInput';
+import { paragraphToWhitespace } from './richtext/utils/richTextHelpers';
 
-type EditorState = {
-  wordCount: number;
-  copiedValue: {
-    text: string | null;
-    seed: number | null;
-  }
-  isTooLong?: boolean;
-}
-type EditorProps = {
-  fieldValue: string;
-  derivedFrom: string;
-  onUpdateField: (string: string) => any;
-  maxLength: number;
-  maxCharLength: number;
-  editable: boolean;
-  fieldLocation: string;
-  fieldName: string;
+
+interface RichTextEditorProps {
+  value: string;
+  onUpdate: (str: string) => void;
   config: EditorConfig;
-  placeholder: string;
-  isDesired?: boolean;
-  isRequired?: boolean;
+  label?: string;
+  copiedValue?: {
+    text: string,
+    seed: number
+  }
+  shouldAcceptCopiedText: boolean;
 }
 
-export default class EditorField extends React.Component<EditorProps, EditorState> {
-  state: EditorState = {
-    wordCount: 0,
-    copiedValue: {
-      text: null,
-      seed: null
-    }
+export const RichTextEditor = ({ value, onUpdate, config, label, copiedValue, shouldAcceptCopiedText = false }: RichTextEditorProps) => {
+  const schema = createSchema(config);
+
+  const editorEl = useRef<HTMLDivElement>(null);
+  const [editorView, setEditorView] = useState<EditorView | undefined>(
+    undefined
+  );
+
+  useEffect(() => {
+    // Editor view takes an HTML Node therefore this string value needs to be converted into a node by placing in a div
+    const contentNode = document.createElement('div');
+    contentNode.innerHTML = value;
+    const edView = createEditorView(onUpdate, editorEl, contentNode, schema, config);
+    setEditorView(edView);
+  }, []);
+
+  const setContent = (tr: Transaction, node: Node): Transaction => {
+    const sel = new AllSelection(tr.doc);
+    sel.replaceWith(tr, node);
+    return tr;
   };
 
-  componentDidMount() {
-    this.updateWordCount(this.props.fieldValue);
-  }
-
-  updateValueFromCopy = () => {
-    this.props.onUpdateField(this.props.derivedFrom);
-    this.setState({
-      copiedValue: {text: this.props.derivedFrom, seed: Math.random()}
-    });
-  };
-
-  updateWordCount = (text: string) => {
-    const count = text ? getWords(text).length : 0;
-
-    this.setState({
-      wordCount: count,
-      isTooLong: false
-    });
-  };
-
-  updateFieldValue = (value: string) => {
-    if (!isTooLong(value, this.props.maxLength, this.props.maxCharLength)) {
-      this.setState({
-        isTooLong: false
-      });
-
-      this.updateWordCount(value);
-      this.props.onUpdateField(value);
-    } else {
-      this.setState({
-        isTooLong: true
-      });
-    }
-  };
-
-  renderCopyButton = () => {
-    if (this.props.derivedFrom === undefined || !this.props.editable) {
-      return null;
-    }
-
-    return (
-        <button
-          type="button"
-          disabled={!this.props.derivedFrom}
-          className="btn form__label__button"
-          onClick={this.updateValueFromCopy}
-          data-tip="Copy trail text from description"
-          data-place="top"
-        >
-          <i className="icon">edit</i>
-        </button>
-        );
-  };
-
-  renderLimitWarning = () => {
-    return (
-      <div>
-        {this.state.isTooLong
-          ? <span className="form__message__text--error">
-              {' '}(This text is too long: updates will not get saved)
-            </span>
-          : false}
-      </div>
-    );
-  };
-
-  renderField() {
-    const requiresValidation = this.state.wordCount === 0  && this.props.fieldLocation === 'trailText';
-    const hasWarning = requiresValidation && this.props.isDesired;
-    const hasError = requiresValidation && this.props.isRequired;
-    if (!this.props.editable) {
-      if (this.state.wordCount === 0) {
-        return (
-          <div className="details-list__field details-list__empty">
-            {this.props.placeholder}
-          </div>
-        );
-      }
-      return (
-        <div
-          className="details-list__field details-list__field-with-content"
-          dangerouslySetInnerHTML={{ __html: this.props.fieldValue }}
-        />
+  useEffect(() => {
+    if (copiedValue.text && editorView && shouldAcceptCopiedText){
+      const parser = parseDOM.fromSchema(schema);
+      const domParser = new DOMParser();
+      const dom = domParser.parseFromString("", "text/html").body;
+      const doc = parser.parse(dom);
+      // Empty the text content of the derived field (trail text)
+      editorView.dispatch(
+          setContent(editorView.state.tr, doc)
       );
+      // Add the new content to that field
+      const valueWithParagraphsAsHardBreaks = paragraphToWhitespace(copiedValue.text);
+      editorView.pasteHTML(valueWithParagraphsAsHardBreaks);
     }
+  }, [copiedValue.text, copiedValue.seed]);
 
-    return (
-      <div>
+  return (
+    <>
+      {label && (
+        <label className="prosemirror__label" htmlFor={label}>
+          <span>{label}</span>
+        </label>
+      )}
+      <div className="prosemirror__input">
         <div
-          className={'form__row editor__row'}
-        >
-          <RichTextInput
-            value={this.props.fieldValue}
-            onUpdate={this.updateFieldValue}
-            copiedValue={this.state.copiedValue}
-            config={this.props.config}
-            shouldAcceptCopiedText={!!this.props.derivedFrom}
-          />
-          {this.renderLimitWarning()}
-        </div>
-        {hasWarning
-          ? <p className="form__message form__message--warning">
-            {RequiredForComposer.warning}
-            </p>
-          : ''}
-        {hasError
-          ? <p className="form__message form__message--error">
-            {RequiredForComposer.error}
-            </p>
-          : ''}
+          className="ProseMirror-example-setup-style"
+          ref={editorEl}
+          data-testid="edit-form-rich-text"
+        />
       </div>
-    );
-  }
-
-  render() {
-    return (
-      <div className="form__row">
-        <div className="form__label__layout">
-          <label className="form__label">{this.props.fieldName}</label>
-          {this.renderCopyButton()}
-        </div>
-        {this.renderField()}
-      </div>
-    );
-  }
-}
+    </>
+  );
+};
